chore: Remove integration test runs
After some consideration, we have decided it might be best to disable integration tests as an experiment. We would like to make sure that they are indeed worth the time and money invested, and that they do catch a class of bugs that occur frequently enough to justify this investment. We know that some cases will now be missed, but we will rely on contributors to do local testing of their own to verify that the APIs work as expected. In many cases, we will still maintain terraform to make this as easy as possible.